# apps/react/src/apps/apidocs/openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/auth/github": {
			"get": {
				"operationId": "AuthController_github",
				"parameters": [],
				"responses": {
					"302": {
						"description": "GitHub 로그인 화면으로 보낸다"
					},
					"429": {
						"description": "짧은 시간에 너무 많이 시도했다"
					}
				},
				"tags": [
					"auth"
				]
			}
		},
		"/auth/github/callback": {
			"get": {
				"operationId": "AuthController_callback",
				"parameters": [
					{
						"name": "code",
						"required": true,
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "state",
						"required": true,
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "error",
						"required": true,
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"302": {
						"description": "프론트엔드로 돌아간다 (?admin=signed-in | denied | cancelled)"
					},
					"429": {
						"description": "짧은 시간에 너무 많이 시도했다"
					}
				},
				"tags": [
					"auth"
				]
			}
		},
		"/auth/me": {
			"get": {
				"operationId": "AuthController_me",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminResponse"
								}
							}
						}
					},
					"401": {
						"description": "관리자로 로그인하지 않았다"
					}
				},
				"security": [
					{
						"macfolio_session": []
					}
				],
				"tags": [
					"auth"
				]
			}
		},
...
		"/posts/{slug}/comments": {
			"get": {
				"operationId": "CommentsController_list",
				"parameters": [
					{
						"name": "slug",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "댓글 (오래된 것부터). mine: 이 브라우저가 쓴 댓글. 방문자·IP 해시는 담지 않는다"
					}
				},
				"tags": [
					"comments"
				]
			},
			"post": {
				"operationId": "CommentsController_create",
				"parameters": [
					{
						"name": "slug",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"example": {
									"body": "잘 봤어요!"
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "쓴 댓글. 처음 쓰는 브라우저에는 방문자 쿠키를 준다"
					},
					"400": {
						"description": "입력 규칙을 어겼다 (내용 1~500자)"
					},
					"429": {
						"description": "짧은 시간에 너무 많이 썼다"
					}
				},
				"tags": [
					"comments"
				]
			}
		},
		"/comments/{id}": {
			"delete": {
				"operationId": "CommentsController_remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "지웠다"
					},
					"403": {
						"description": "다른 브라우저에서 쓴 댓글이다"
					},
					"404": {
						"description": "댓글이 없다"
					}
				},
				"tags": [
					"comments"
				]
			}
		},
...
	},
	"info": {
		"title": "MacFolio API",
		"description": "메시지(피드백), 블로그(관리자 편집), 메일",
		"version": "0.1.0",
		"contact": {}
	},
...
	"servers": [],
	"components": {
		"securitySchemes": {
			"cookie": {
				"type": "apiKey",
				"in": "cookie",
				"name": "macfolio_session"
			}
		},
		"schemas": {
...
			"AdminResponse": {
				"type": "object",
				"properties": {
					"login": {
						"type": "string",
						"example": "hyeoniverse"
					},
					"signedInAt": {
						"type": "string",
						"example": "2026-10-02T10:03:00.000Z",
						"description": "이 세션으로 로그인한 때 (ISO 8601)"
					},
					"expiresAt": {
						"type": "string",
						"example": "2026-10-02T22:03:00.000Z",
						"description": "이 세션이 끝나는 때 (ISO 8601)"
					}
				},
				"required": [
					"login",
					"signedInAt",
					"expiresAt"
				]
			}
		}
	}
}